+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Throw error if invalid Payload on `CustomerGateway.create`
+* Allow payee ID to be passed in options params for transaction create
 
 ## 2.10.0
 * Bank Account verifications API